Add: Steam folder + modify pdf-mix-tool
Add the Steam folder icon for each set of folder icons/colors.
Because the folder had to be located in the color folder as well as in
the src/, a physical link was made to avoid trouble if the icons need to
be modified.

Add modification to the pdf-mix-tool icon, this make it more uniform
with the other icons.