FIX #58 - Authy
Add Authy icon as requested in #58.

Modify the README.md. I noticed that the snaps icons are not following
the theme. This is because their .desktop is pointing to somewhere else,
but not to the usual places. An alternative is to copy the .desktop from
/var/lib/snapd/desktop/applications/name-of-snap-application.destop into
the local $HOME/.local/share/applications/ and modify the Icon= line to
point to the right one. This allow therefore a backup (original
.desktop) as well as avoid making change to these file, as each update
of the snaps will probably override these .desktop file, but not the
copy in the home folder.