FIX: #30 Non-apple alternative
This commit aims to resolve issue #30
It add 2 "alternative" icons, for softwarecenter and file-manager.
To use them, I added a parameter (-a or --alternative) in the install.sh. When specified,
it will call a function to change the name of the icons and use the
"alternative" one instead of the "apple" icons during installation.
At the end of the installation, it reset the icons to their default
value.